# Improve Spark test config