--- a/caps-spark/src/test/scala/org/opencypher/caps/test/TestSparkSession.scala
+++ b/caps-spark/src/test/scala/org/opencypher/caps/test/TestSparkSession.scala
@@ -44,14 +44,15 @@ object TestSparkSession {
     val conf = new SparkConf(true)
 
     conf.set("spark.sql.codegen.wholeStage", "true")
-
-    // conf.set("spark.rdd.compress", "true")
-
     conf.set("spark.serializer", classOf[KryoSerializer].getCanonicalName)
     conf.set("spark.kryo.registrator", classOf[CypherKryoRegistrar].getCanonicalName)
     conf.set("spark.kryo.unsafe", "true")
     conf.set("spark.kryo.referenceTracking","false")
     conf.set("spark.kryo.registrationRequired", "true")
+    conf.set("spark.sql.shuffle.partitions", "1")
+    conf.set("spark.sql.inMemoryColumnarStorage.compressed", "false")
+    conf.set("spark.submit.deployMode", "client")
+
     //
     // If this is slow, you might be hitting: http://bugs.java.com/view_bug.do?bug_id=8077102
     //
@@ -66,6 +67,3 @@ object TestSparkSession {
     session
   }
 }
-
-
-