--- a/pom.xml
+++ b/pom.xml
@@ -40,7 +40,7 @@
     <plugin.maven-resources.version>2.7</plugin.maven-resources.version>
     <plugin.maven-scala.version>3.2.1</plugin.maven-scala.version>
     <plugin.maven-scalastyle.version>0.8.0</plugin.maven-scalastyle.version>
-    <plugin.maven-scalatest.version>1.0</plugin.maven-scalatest.version>
+    <plugin.maven-scalatest.version>2.0.0</plugin.maven-scalatest.version>
     <plugin.maven-shade.version>2.4.3</plugin.maven-shade.version>
     <plugin.maven-surefire.version>2.20.1</plugin.maven-surefire.version>
 
@@ -227,6 +227,7 @@
           <reportsDirectory>${project.build.directory}/surefire-reports</reportsDirectory>
           <junitxml>.</junitxml>
           <filereports>WDF TestSuite.txt</filereports>
+          <parallel>true</parallel>
         </configuration>
         <executions>
           <execution>