@pstutz 2018-02-12
Disabled parallel test execution. The improvement is mainly due to the Spark config change, and the extra 10 seconds are not worth getting test results out of order in the log.